# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FULL SWAP API - Real-time AI video transformation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a USDT payment (optionally supply tx hash; auto-scans blockchain if omitted)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerification"
  /payments/status/{invoiceId}:
    get:
      operationId: getInvoiceStatus
      tags: [payments]
      summary: Poll on-chain status of an invoice (auto-verifies if confirmed)
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice blockchain status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceStatus"
...
components:
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        userId:
          type: integer
        minutes:
          type: integer
        amountUsdt:
          type: number
        status:
          type: string
          enum: [pending, paid, expired]
        walletAddress:
          type: string
        walletNetwork:
          type: string
          nullable: true
        allWallets:
          type: array
          nullable: true
          items:
            type: object
            properties:
              address:
                type: string
              network:
                type: string
            required: [address, network]
        txHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, minutes, amountUsdt, status, walletAddress, createdAt]
    VerifyPaymentBody:
      type: object
      properties:
        invoiceId:
          type: string
        txHash:
          type: string
      required: [invoiceId]
    InvoiceStatus:
      type: object
      properties:
        invoiceId:
          type: string
        status:
          type: string
        confirmed:
          type: boolean
        confirmations:
          type: integer
          nullable: true
        txHash:
          type: string
          nullable: true
        minutesAdded:
          type: integer
        message:
          type: string
      required: [invoiceId, status, confirmed, minutesAdded, message]
    PaymentVerification:
      type: object
      properties:
        verified:
          type: boolean
        invoice:
          $ref: "#/components/schemas/Invoice"
        minutesAdded:
          type: integer
      required: [verified, invoice, minutesAdded]
...